Fixing schema/hibernate mapping issues
